# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: BTS Member Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/tracks/{trackId}/modules:
    get:
      operationId: adminListModules
      tags: [admin-content]
      summary: List modules within a track
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminModule"
  /admin/modules:
    post:
      operationId: adminCreateModule
      tags: [admin-content]
      summary: Create a new module
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateModule"
      responses:
        "201":
          description: Module created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminModuleDetail"
  /admin/modules/{id}:
    put:
      operationId: adminUpdateModule
      tags: [admin-content]
      summary: Update a module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateModule"
      responses:
        "200":
          description: Module updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminModuleDetail"
    delete:
      operationId: adminDeleteModule
      tags: [admin-content]
      summary: Delete a module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Module deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminModuleDetail"
  /admin/modules/reorder:
    patch:
      operationId: adminReorderModules
      tags: [admin-content]
      summary: Reorder modules within a track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderRequest"
      responses:
        "200":
          description: Modules reordered
  /admin/modules/{id}/move:
    patch:
      operationId: adminMoveModule
      tags: [admin-content]
      summary: Move a module to a different track
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetTrackId:
                  type: integer
              required: [targetTrackId]
      responses:
        "200":
          description: Module moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminModuleDetail"
...
  /admin/lessons/{lessonId}/resources/upload-url:
    post:
      operationId: adminRequestResourceUploadUrl
      tags: [admin-content]
      summary: Request upload URL for a lesson resource
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceUploadRequest"
      responses:
        "200":
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
  /admin/lessons/{lessonId}/resources:
    get:
      operationId: adminListLessonResources
      tags: [admin-content]
      summary: List resources for a lesson
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LessonResource"
    post:
      operationId: adminCreateLessonResource
      tags: [admin-content]
      summary: Create a lesson resource record after upload
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonResource"
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonResource"
  /admin/lessons/{lessonId}/resources/reorder:
    patch:
      operationId: adminReorderLessonResources
      tags: [admin-content]
      summary: Reorder lesson resources
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderRequest"
      responses:
        "200":
          description: Resources reordered
  /admin/resources/{id}:
    delete:
      operationId: adminDeleteResource
      tags: [admin-content]
      summary: Delete a lesson resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resource deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonResource"
...
  /lessons/{lessonId}/resources/{resourceId}/download:
    get:
      operationId: downloadLessonResource
      tags: [lessons]
      summary: Download a lesson resource (entitlement-checked)
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
        - name: resourceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File download
        "403":
          description: Access denied
components:
  schemas:
...
    ReorderRequest:
      type: object
      properties:
        orders:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              sortOrder:
                type: integer
            required: [id, sortOrder]
      required: [orders]
    AdminModule:
      type: object
      properties:
        id:
          type: integer
        trackId:
          type: integer
        title:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
        lessonCount:
          type: integer
      required: [id, trackId, title, description, sortOrder, lessonCount]
    AdminModuleDetail:
      type: object
      properties:
        id:
          type: integer
        trackId:
          type: integer
        title:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
      required: [id, trackId, title, description, sortOrder]
    AdminCreateModule:
      type: object
      properties:
        trackId:
          type: integer
        title:
          type: string
        description:
          type: string
      required: [trackId, title, description]
    AdminUpdateModule:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
...
    LessonResource:
      type: object
      properties:
        id:
          type: integer
        lessonId:
          type: integer
        fileName:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: integer
        fileType:
          type: string
        sortOrder:
          type: integer
        downloadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, lessonId, fileName, fileUrl, fileSize, fileType, sortOrder, downloadCount, createdAt]
    CreateLessonResource:
      type: object
      properties:
        fileName:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: integer
        fileType:
          type: string
      required: [fileName, fileUrl, fileType]
    ResourceUploadRequest:
      type: object
      properties:
        fileName:
          type: string
        fileSize:
          type: integer
        fileType:
          type: string
      required: [fileName, fileSize, fileType]
    UploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required: [uploadURL, objectPath]
...