# lib/api-zod/src/generated/types/lessonResource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BTS Member Portal API
 * OpenAPI spec version: 0.2.0
 */

export interface LessonResource {
  id: number;
  lessonId: number;
  fileName: string;
  fileUrl: string;
  fileSize: number;
  fileType: string;
  sortOrder: number;
  downloadCount: number;
  createdAt: Date;
}